Added getting, saving, and reading of hash id.
Added is_saved_id to check if a hash ID has been saved to the current directory.

Added save_hash_id that takes in an id and saves it to a file called hash.txt. Overwrites the current file if it exists.

Added get_hash function to either get a new hash ID or to open the current saved hash and return it.

- Notes
-- Should probably split get_hash into two separate functions. Eventually, we'll want users to be able to enter in commands to change the current hash number (for whatever reason).

-- Idk